dsp-kernel: Handle UAF scenario in put_args
 Currently, the DSP updates header buffers with unused DMA handle fds.
In the put_args section, if any DMA handle FDs are present in the
header buffer, the corresponding map is freed. However, since the
header buffer is exposed to users in unsigned PD, users can update
invalid FDs. If this invalid FD matches with any FD that is already
in use, it could lead to a use-after-free (UAF) vulnerability.
As a solution,add DMA handle references for DMA FDs, and the map for
the FD will be freed only when a reference is found.

Change-Id: Id4115e9be33cd69944a8cd04e8b336280b8b53be
Signed-off-by: Minghao Xue <quic_mingxue@quicinc.com>
(cherry picked from commit c50f6839522acc44e019b5e56d336c5b3c32878d)