platform-kernel:virtio feature check in rsm probe
Changes to perform the NSP Sharing feature check during probe and not
proceed with the virt queue init and other allocations. This prevents
a crash if the PVM driver is not loaded while the LA GVM is loaded.
Also include a fix to handle when the txbuf limit is reached.

Change-Id: I30102b35506349c1b370f16e58658dbb066483bc
(cherry picked from commit e64d8f77c01e74fc386e926ff1a357ba15f431b2)